ONLINE RESERVATION SYSTEM
Task- 1 ONLINE RESERVATION SYSTEM 

From completing this task of creating an Online Reservation System, I have learned how to design and develop a functional system using Java programming language and a central database. I have gained experience in creating different modules such as a login form, reservation system, and cancellation form, which allowed users to interact with the system and perform the necessary actions. Additionally, I have learned the importance of creating a user-friendly interface that is easy to navigate and ensuring that data is secure and reliable. Overall, this task has helped me develop my skills in Java programming and database management, and I feel more confident in creating similar systems in the future.